feat: refactor package structure and consolidate tests while preserving behavior
